refactor: create a better structure for the sources
